@@ -10,8 +10,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
-        "illuminate/support": "~4"
+        "php": ">=5.4.0",
+        "illuminate/support": "~5.0"
     },
     "require-dev": {
         "mockery/mockery": "0.9.*"
